Map screen refactoring
