+++ StoryMap/StoryMap/Supporting files/Info.plist
@@ -62,5 +62,7 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
+	<key>UIUserInterfaceStyle</key>
+	<string>LIGHT</string>
 </dict>
 </plist>
